Renamed package structure Got rid of duplicate test folders within each folder and moved all tests to one test folder
