+++ src/server/ShirtCredentialsManager.java
@@ -1,4 +1,4 @@
-package model.server;
+package server;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
