+++ src/views/Business.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.paint.Stop?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="498.0" prefWidth="885.0" style="-fx-background-color: #6495ED;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.BusinessCodeBehind">
+<AnchorPane prefHeight="498.0" prefWidth="885.0" style="-fx-background-color: #6495ED;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="views.BusinessCodeBehind">
     <children>
       <Label layoutX="274.0" layoutY="14.0" text="TeeGenius">
          <font>
