+++ src/views/Login.fxml
@@ -2,7 +2,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane prefHeight="200.0" prefWidth="320.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.LoginController">
+<AnchorPane prefHeight="200.0" prefWidth="320.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="views.LoginController">
     <children>
     	<Label layoutX="27.0" layoutY="10.0" text="Account Type:"/>
         <ComboBox fx:id="accountTypeComboBox" layoutX="112.0" layoutY="10.0" prefWidth="150.0"/>
