+++ src/views/ShirtCreatorCodeBehind.java
@@ -1,4 +1,4 @@
-package view;
+package views;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
